logging: Always pad timestamp to same length
This fixes inconsistent looking timestamps in the logging output by always
padding the microsecond component to 6 digits

+++ libs/pilight/core/log.c
@@ -185,7 +185,7 @@ void logprintf(int prio, const char *format_str, ...) {
 		if((gmtime_r(&tv.tv_sec, &tm)) != 0) {
 #endif
 			strftime(fmt, sizeof(fmt), "%b %d %H:%M:%S", &tm);
-			snprintf(buf, sizeof(buf), "%s:%03u", fmt, (unsigned int)tv.tv_usec);
+			snprintf(buf, sizeof(buf), "%s:%06ld", fmt, tv.tv_usec);
 		}
 		pos += sprintf(line, "[%22.22s] %s: ", buf, progname);
